validations added on price unit-- NOT WORKING ON PARTIAL

+++ src/price-unit/price-unit.controller.ts
@@ -6,12 +6,17 @@ import {
   Delete,
   Param,
   Body,
+  UsePipes,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 
 import { PriceUnitService } from './price-unit.service';
 import { PriceUnitDTO } from './price-unit.dto';
+import { ValidationPipe } from '../shared/validation.pipe';
+var validate = require('uuid-validate');
 
-@Controller('unit')
+@Controller('units')
 export class PriceUnitController {
   constructor(private priceUnitService: PriceUnitService) {}
 
@@ -22,10 +27,15 @@ export class PriceUnitController {
 
   @Get(':id')
   getUnitById(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exceptio
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.priceUnitService.read(id);
   }
 
   @Post()
+  @UsePipes(new ValidationPipe())
   createPriceUnit(@Body() data: Partial<PriceUnitDTO>) {
     return this.priceUnitService.create(data);
   }
@@ -35,11 +45,19 @@ export class PriceUnitController {
     @Param('id') id: string,
     @Body() data: Partial<PriceUnitDTO>,
   ) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.priceUnitService.update(id, data);
   }
 
   @Delete(':id')
   deletePriceUnit(@Param('id') id: string) {
+    if (!validate(id)) {
+      //If Id is not valid then throw exception
+      throw new HttpException('Not Valid id', HttpStatus.BAD_REQUEST);
+    }
     return this.priceUnitService.destroy(id);
   }
 }
